requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# MongoDB Database
pymongo==4.6.0

# IPFS Client
ipfshttpclient==0.8.0a2

# Cryptography
cryptography==41.0.7
pycryptodome==3.19.0

# JWT Authentication
PyJWT==2.8.0

# HTTP Requests
requests==2.31.0

# Web3 and Ethereum
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Environment Variables
python-dotenv==1.0.0

# Date/Time Utilities
python-dateutil==2.8.2

# Data Validation
marshmallow==3.20.1

# Password Hashing
bcrypt==4.1.2

# CORS Support
flask-cors==4.0.0

# Logging
python-json-logger==2.0.7

# File Type Detection
python-magic==0.4.27

# Base58 Encoding (for IPFS CIDs)
base58==2.1.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Optional: ML/AI Libraries (Comment out if not using ZKML features)
# tensorflow==2.15.0
# onnxruntime==1.16.3
# numpy==1.24.3
# scikit-learn==1.3.2

# Document Processing
PyMuPDF==1.23.5
python-docx==1.1.0
Pillow==10.1.0

# Production Server (for deployment)
gunicorn==21.2.0
gevent==23.9.1
